Version 1.0
This is the first definitive version that is bug-free and has all the necessary stuff. Over the next few days, I will test it and then update what needs to be updated. I am also looking forward to making this more accessible to new people who have never touched neovim.